Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> I'm not sure about both the name and the "HaikuExt" namespace. "libNetworkKit/libnetworkkit" seems like a better name overall, and what is wrong with BPrivate?

I think we should avoid BPrivate because it is a namespace that already exists in our official header files and libraries. Concretely, BHttpTime already lives in the BPrivate namespace. I want to avoid odd compiling and/or linking errors for the users.

In my thinking, I have a preference not to use any namespace that we may want to use in the future. I suppose HaikuExt may fit that bill, but so will libNetworkKit (or libNetServices?), so I am open to change if this becomes an issue.

Patch-set: 1
Reviewer: Gerrit User 1000012 <1000012@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>