Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1: Code-Review-1
> > 
> > I'm not sure about both the name and the "HaikuExt" namespace. "libNetworkKit/libnetworkkit" seems like a better name overall, and what is wrong with BPrivate?
> 
> I think we should avoid BPrivate because it is a namespace that already exists in our official header files and libraries. Concretely, BHttpTime already lives in the BPrivate namespace. I want to avoid odd compiling and/or linking errors for the users.
> 
> In my thinking, I have a preference not to use any namespace that we may want to use in the future. I suppose HaikuExt may fit that bill, but so will libNetworkKit (or libNetServices?), so I am open to change if this becomes an issue.

I prefer BPrivate as well. This is what it is designed for: isolating private and work-in-progress APIs in a single namespace.

Also, we just can't introduce namespaces with arbitrary names. The convention (though it is an unwritten one) is that all names starting with B + another uppercase letter are reserved for us (the operating system implementing the BeAPI), and that everything else is for applications. So, if you really don't want to use BPrivate, try to at least stick to that convention.

If you fear of conflicts inside BPrivate, you can use a nested namespace such as BPrivate::Network.

To do things as we do in libshared, the .h files would also include an "using namespace BPrivate::Network;", so that this namespace remains largely an internal application detail. It may not be possible here because of the already existing classes in libbnetapi, however.

(also I'm not particularly fan of the idea of including "kit" in the library names, I'd keep the convention used by BeOS for libnetwork, libgame, libtranslation, libtracker. and we need a better name than "service kit" or "network services", but that can be solved later, I guess).

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>