Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> I am once again not sure about the terminology. To me as a native English speaker, "medium upgrade" sounds like "medium sized upgrade", as if one is at a restaurant...

It would be a large if it would include EFI.

The other commands seem to be verb - noun (add repo, resolve dependencies), so it is inconsistent with that. Though upgrade-medium sounds even more odd to me.

Again, my argument for including `medium` is that the upgrade does not just expect a directory with packages, but it expects something that is laid out like an installation medium, so I am still in favor of keeping the word.

What about `upgrade from medium`?

Patch-set: 4
Reviewer: Gerrit User 1000012 <1000012@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>