Update patch set 4
Patch Set 4:

> Patch Set 4:
> It would be a large if it would include EFI.
> 
> The other commands seem to be verb - noun (add repo, resolve dependencies), so it is inconsistent with that. Though upgrade-medium sounds even more odd to me.
> 
> Again, my argument for including `medium` is that the upgrade does not just expect a directory with packages, but it expects something that is laid out like an installation medium, so I am still in favor of keeping the word.
> 
> What about `upgrade from medium`?

I'd be averse to using the word 'medium' here at all to be honest. Most users aren't going to be thinking about the class of all possible installation media as a whole, but rather their USB stick or ISO or what have you. Hence the suggestion of making it an option on pkgman upgrade, i.e. --source /path/to/...

Patch-set: 4