Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Can we use "haiku-os.org" without a scheme? That might work. We should, however, change pkgman et al. to at least display "identifier" instead of "url" so users are not confused still.

pkgman and everything already shows "identifier:". This part of the change was done before beta2 already. Some code in haikudepot still relied on the field being an uri of some form (which is why a tag: uri is proposed here). I don't understand why you are so against using the "tag:" scheme, which provides some way to format the URIs and make sure there are no accidental conflicts.

Yes, for Haiku repositories we can use whatever we want. But, we also need to provide some naming convention, so that other repositories from 3rd party can generate an unique identifier for themselves. The "tag:" URI scheme does this for free: everyone should use their own domain name and a date at which they own it. Then they can organize themselves as they wish inside that namespace. Another option is to use the "uuid:" URI scheme where no manual intervention is needed (you just generate an uuid and it will very probably be unique). And yet another is to use http: and match the repo URI (what we have done until now, creating some confusion when we move or mirror repos). And if later we find something better, we can wrap that new thing into a different URI scheme and it will never clash with existing repos.

If we don't use standard URI schemes, we need to provide some other way for other people to define identifiers for their repos in an unique way. It creates more work for us (we have to specify how that will work) and I don't think we can come up with something better than what is done in "tag:" already, unless we go with something like "email someone at Haiku and they can check if the name you want is currently available".

As a final note, these identifiers are not supposed to be manipulated directly by users most of the time. So I see no point in discussing the exact format in that aspect. "it looks nice" isn't really an important feature here. You do not use this identifier when adding a repository, it is only used internally to check if what's stored at some URL is a mirror of some other URL (and even then, it is not sufficient on its own to guarantee that, some signature system should additionally be used).

So, if "tag:" is not making people happy, I suggest we simply go with "uuid:" instead.

Patch-set: 2
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>