Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > > What about `upgrade from medium`?
> > 
> > I'd be averse to using the word 'medium' here at all to be honest. Most users aren't going to be thinking about the class of all possible installation media as a whole, but rather their USB stick or ISO or what have you. Hence the suggestion of making it an option on pkgman upgrade, i.e. --source /path/to/...
> 
> Yes, it seems "upgrade --from /path/..." makes more sense. And we don't even need to name the "from" thing more specifically, which is one way to solve the problem.

Not sure whether there is a misunderstanding here, but pkgman does not have `upgrade`. It has `update` and `full-sync`. 

If the idea is to make it an option for the update command, then am I understanding correctly that if the given parameter is a path, that we run the 'is it a medium' logic and if so proceed with installing it as if it was a medium?

Patch-set: 4
Reviewer: Gerrit User 1000012 <1000012@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>