Update patch set 1
Patch Set 1:

To be clear: this should not be merged into the master branch, but in the beta3 branch when we create it. Because once we merge this we may need everyone to re-add the repos, and it's better to do this once (when updating to beta3) than twice (when merging this, and again when creating the beta3 repos).

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>