Update patch set 1
Patch Set 1:

> @apl: if we make this change, will HaikuDepotServer and HaikuDepot handle it?

Hello Niels; HDS would pick-up the change in "repo.info" and start using it after the next repository refresh.  I can't see any problems there.  HDS and HaikuDepot both just expect a string for the identifier so both will handle the new "tag:..." format.

Mid 2020 I updated HDS and HaikuDepot to be able to handle "additional identifiers".  Additional identifiers are able to be configured against a repository source in HDS (editable in the UI as a root user) so that a (relatively recent) HaikuDepot client is able to work with either the current identifier as published in the "repo.info" or an older identifier.  This may make the transition a bit easier with regard to HDS and HaikuDepot.

Patch-set: 1