Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> Linux implementation: https://github.com/esmil/musl/blob/master/src/misc/ioctl.c
> 
> It passes a pointer to arguments in syscall instead of argument values.

This implementation in musl passes only a pointer (argument 3). It does not implement ioctl with 4 parameters.

The main difference we have with other OS is that they have a standard encoding of the second parameter. For example in FreeBSD: https://www.freebsd.org/cgi/man.cgi?query=ioctl&sektion=2 :

> An ioctl() request has encoded in it whether the argument is an "in" argument or "out" argument, and the size of the argument argp in bytes. Macros and defines used in specifying an ioctl request are located in the file <sys/ioctl.h>.

A similar description is used in Linux manpage.

We don't do that (and can't do that because of BeOS compatibility), and we instead have a 4-argument version of ioctl, where the 4th argument is the size of the thing pointed by the 3rd. In our case, drivers are free to assign ioctl request values however they want, and the ioctl function knows nothing about it.

>> at worst, it will just crash

> It is not possible under regular conditions

Not on x86 with gcc, but we can't assume anything about how varargs are implemented on different architectures and compilers. Calling va_arg when no more arguments are left is undefined behavior in the C specification (source: https://gist.github.com/Earnestly/7c903f481ff9d29a3dd1) and we should not rely on implementation details of it.

Patch-set: 4
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>