Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Does it cause add-on memory leak? If so, TODO should be added before proper fix.

No, a single call remains to UnloadIfPossible() in the CatalogAddOn destructor and that is enough to free the memory.

Before the actual add-on code was loaded/unloaded on-demand, now it's loaded once at start and unloaded when quitting the app. Which I think makes more sense, there is a single add-on (to load plaintest catalog files), it is small, and loading it from disk everytime we're scanning for catalogs sounds like a bad idea for performance.

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>