Update patch set 1
Patch Set 1:

Having a separate EFI app will only move the complexity elsewhere: currently in the hands of the user trying to setup Haiku on an affected machine. Possibly later on in Installer when it does EFI bootloader installation. I think it's better to keep the bootloader an all-in-one thing.

However, it seems this patch does not change anything according to the bugreport updates...

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>