Update patch set 1
Patch Set 1:

Should we use a syslog trace in releases, and a panic in nightly builds?

What's the risk in continuing here? Will the affected app crash "safely" or will it just continue running but possibly behave strangely? The latter could result in frustration for the app developer (who might not be watching the syslog)

If the app cleanly crashes or if the situation can be solved, there isn't really a need for a panic

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>