Update patch set 1
Patch Set 1:

I am not going to work on this proposed change before beta 3 lands, as it makes so many changes to the class hierarchy so that it will lead to a mess of #ifdefs. 

Having said that, I am unsure whether to pursue this even further. I do not really see the advantage to this approach versus the BUrlContext/BUrlProtocolRoster approach, but I am open for any further thoughts on why this may be good.

Patch-set: 1