Update patch set 1
Patch Set 1:

> Destruction of BUrlSession when there are still requests running is
> undefined. Mitigations were considered:

Under the older API, the "listener" received all of the lifecycle events from the URL's operation.  The method `RequestCompleted` was a safe place to clean up any BDataIO objects that needed to be cleaned up because the listener owned the BDataIO object.  With the listener now detached from the BDataIO, it seems that clean-up is a more difficult because the listener does not have the BDataIO to act on.  Was the rationale of splitting the listener from the BDataIO convenience to avoid having to create a listener?

> A solution might be to use the added WaitForCompletion() to ensure that
> BUrlSession won't be destroyed before the requests themselves stop.
...
> Make BUrlSession a BReferenceable, then every request may carry a reference
> to keep the session alive.

That might not be a good idea because if a request is blocking for a long time then it could cause a quasi leak.

Perhaps a better approach here might have been to create a "standard simple-case listener class" that can wrap a BDataIO instead of divorcing the listener from the data?

Patch-set: 1
CC: Gerrit User 1000018 <1000018@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>