Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Destruction of BUrlSession when there are still requests running is
> > undefined. Mitigations were considered:
> 
> Under the older API, the "listener" received all of the lifecycle events from the URL's operation.  The method `RequestCompleted` was a safe place to clean up any BDataIO objects that needed to be cleaned up because the listener owned the BDataIO object.  With the listener now detached from the BDataIO, it seems that clean-up is a more difficult because the listener does not have the BDataIO to act on.  Was the rationale of splitting the listener from the BDataIO convenience to avoid having to create a listener?
> 
> > A solution might be to use the added WaitForCompletion() to ensure that
> > BUrlSession won't be destroyed before the requests themselves stop.
> ...
> > Make BUrlSession a BReferenceable, then every request may carry a reference
> > to keep the session alive.
> 
> That might not be a good idea because if a request is blocking for a long time then it could cause a quasi leak.
> 
> Perhaps a better approach here might have been to create a "standard simple-case listener class" that can wrap a BDataIO instead of divorcing the listener from the data?

I am not sure whether the listener is such a good idea. I think the concept of a listener does not fit nicely in Haiku's model where all intra-thread communication is done using the Looper/Handler/Message framework. Now that the data reading-writing function of the listener is superfluous, I am wondering whether it is not better to use the standard messaging framework to inform a BMessenger target that something has changed. 

So the logic would be:

    // BMyWindow::_StartRequest()
    fRequest = BUrlProtocolRoster::MakeRequest(...);
    fRequest->NotifyAll(this); // Set notifications
    fRequest->Run(); // Start request

    // BMyWindow::MessageReceived(BMessage *msg)
    case B_URL_PROTOCOL_UPLOAD_PROGRESS:
        off_t bytesSent;
        off_t bytesTotal;
        message->FindInt64("url:bytesSent", &bytesSent);
        message->FindInt64("url:bytesTotal", &bytesTotal);
        _UpdateUploadProgress(bytesSent, bytestotal);
        break;

    case B_URL_PROTOCOL_REQUEST_COMPLETED:
        delete fRequest;
        fRequest = NULL;
        _DoSomethingWithTheResult();
        break;

The messages will be enhanced by something like a unique identifier for the BUrlRequest, for loopers that need to run multiple requests simultaneously. This could be generated by the BUrlRequest class itself, and initially can be the thread_id of the executing thread.

The upside IMO is that we can more clearly tell the user that the BUrlRequest is not thread-safe, and that the user is responsible to clearly have one owning thread. It will introduce the semantics that the asynchronous framework depends on having a BLooper, but then again, that's Haiku for you.

Patch-set: 1