Update patch set 1
Patch Set 1:

> I am not sure whether the listener is such a good idea. I think the concept of a listener does not fit nicely in Haiku's model where all intra-thread communication is done using the Looper/Handler/Message framework. Now that the data reading-writing function of the listener is superfluous, I am wondering whether it is not better to use the standard messaging framework to inform a BMessenger target that something has changed. 

If you want to do this you can use https://git.haiku-os.org/haiku/tree/src/kits/network/libnetservices/UrlProtocolDispatchingListener.cpp

> The messages will be enhanced by something like a unique identifier for the BUrlRequest, for loopers that need to run multiple requests simultaneously. This could be generated by the BUrlRequest class itself, and initially can be the thread_id of the executing thread.

The messages get a pointer to the request they can use for that.

> The upside IMO is that we can more clearly tell the user that the BUrlRequest is not thread-safe, and that the user is responsible to clearly have one owning thread. It will introduce the semantics that the asynchronous framework depends on having a BLooper, but then again, that's Haiku for you.

It is optional, but already possible in the current API. You can use it in two ways:
- Either you set a simple listener directly, and your code gets called in the HTTP thread
- Or, you use an asynchronous listener (https://git.haiku-os.org/haiku/tree/src/kits/network/libnetservices/UrlProtocolAsynchronousListener.cpp), which you attach to one of your own BHandlers. This then takes care of setting up a dispatching listener in the HTTP thread, and converting the messages back into callbacks, now running inside your own thread

Having a separate thread and looper to handle things is not always appropriate. Sometimes you want to just wait for one specific event (say, download complete) and doing that in an asynchronous listener is fine. But sometimes you have more complex things going on (say, streaming a video from youtube and you want to enable seeking). In that case, having two threads to worry about (the HTTP thread running the request, and your own looper handling the data and controlling it) makes things a lot more complicated than they need to.

Also, in haikuwebkit, it is already annoying to have one thread started per request, and WebKit own threads are not really usable as generic BLoopers, so it would probably need a second thread to handle the messages and somehow inject them back into webkit APIs. That would make things more complicated to handle there as well.

Patch-set: 1