Update patch set 1
Patch Set 1:

My main reservation with the current API is that it is not very clear in communicating how the user should use it. The library is internally multithreaded, with the intention of being used in an externally multithreaded environment, meaning we should do our best to make sure the users of the API don't get into accidents.

IMO, this means in order of preference:
1. Make misuse impossible.
2. Have the safest possible defaults and make the user of the API explicitly opt-in to more dangerous behavior.
3. Have runtime checks that spot bad behavior and block/mitigate it.
4. Have a consistent design of the API so that users can use their previous knowledge to do the right thing.
5. If we cannot do any of those things, then make sure the documentation covers it clearly.

My main issue with the listener is that it will have us fall back to #5. The issue I see is that there are a few threads involved:
Thread A: owns the BUrlRequest
Thread B: the network execution thread, which does not own any data and does not do any locking. It exists outside of the user's control.
Thread C (potentially): the API makes it possible that the listener is actually owned by another thread.

The API user will need to make sure to make a full mental model of how their threads interact, how to deal with lifetimes and how to deal with synchronization/locking. They will also need to understand the inner working of the services kit API and understand that there is another thread involved that they do not create or control themselves, but that does execute code that impacts their other own thread(s).

In this scenario, if I run through the list on how to make a good API, these constraints are very badly communicated. Going through the list: misuse is absolutely possible and arguably by default because it is very easy to write multithreaded code that works 99.5% of the cases, but gives odd race-condition bugs especially in failure paths that are not encountered often (#1). In the current API, the constructor of the request asks for a listener, so from an ergonomic perspective the functionality is not explicitly opt-in (#2). There are no runtime checks, and I don't see how they are possible (#3). The listener model is not based on any other exiting Haiku functionality (#4), also developers coming from the Android world may be mislead, because on Android listeners on Activities (i.e. UI) are by default executed on the UI thread instead of the worker thread. So that means that our strategy for the best possible use of the API is through documentation (#5), which should really be the last resort.

Maybe the listener API is the only way to get all the functionality we want, but I do want to think about alternatives and/or ergonomics that make its use better.

> Patch Set 1: 
> Also, in haikuwebkit, it is already annoying to have one thread started per request, and WebKit own threads are not really usable as generic BLoopers, so it would probably need a second thread to handle the messages and somehow inject them back into webkit APIs. That would make things more complicated to handle there as well.

I am on the fence on whether haikuwebkit is a good target for the services kit. It is good when our kit can support it, but the downsides are that (1) it is in essence not really Haiku-native code and (2) it probably benefits from more specialized HTTP network code (especially for HTTP2/3), than the general purpose toolkit we are trying to give our users.

Patch-set: 1