efi/riscv64: Fix NumberOfRvaAndSizes of PE header
The riscv64 UEFI executable has inconsistent header and therefore cannot
be loaded by the EDK2 implementation of UEFI. Apparently Uboot doesn't
care about this detail but EDK2 validates it.

Specifically NumberOfRvaAndSizes should be the number of data directory
entries. It is set to 6 but there are actually 16.
(All entries below NumberOfRvaAndSizes until section_table)

This was probably a typo 0x6 -> 16 (or 0x10).

This is the check of EDK2:
https://github.com/tianocore/edk2/blob/stable/202011/MdePkg/Library/BasePeCoffLib/BasePeCoff.c#L313-L321

Fixes #16824

Change-Id: I5c0dc080cb1cd191b44e016298656db9edd9ca8d