DriveSetup: fix logic to enable the "change parameters" menu
Just ask the partition if its parameters can be changed, instead of
trying to guess.

Now the menu is always disabled. This is because none of our
partitioning systems actually allow changing any of the parameters of an
already created partition (they don't enable B_DISK_SYSTEM_SUPPORTS_SETTING_PARAMETERS).

Fixes #11641 (but we should implement editing parameters in the intel
partitionning system to set the active partition).