Update patch set 1
Patch Set 1:

(1 comment)

I am a bit worried to see various checks for errors and "return B_NO_MEMORY;" simply removed. Vectors can also fail at inserting memory and will throw exceptions (std::bad_alloc), is there something in place to catch that? Having the whole program just disappear without an error message is not great, so the exceptions should be caught somewhere at least to display an error alert, and ideally to try to continue running as much as possible.

I have put an inline suggestion for a TODO that can now easily be solved, but maybe it makes sense to handle this as a later commit.

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>