Update patch set 1
Patch Set 1:

(1 comment)

I tried the patch and I am now in a situation where the menu item is disabled everywhere. I am a bit confused which functionality we want to show here. I was expecting that I could at least change the type of filesystem on an unformatted partition.

Patch-set: 1
CC: Gerrit User 1000012 <1000012@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>