Update patch set 1
Patch Set 1:

(1 comment)

> I tried the patch and I am now in a situation where the menu item is disabled everywhere. I am a bit confused which functionality we want to show here. I was expecting that I could at least change the type of filesystem on an unformatted partition.

Currently the intel partition map add-on does not have the needed code to edit anything once the partition is created. So the menu is now (correctly) always disabled. I am working on adding the required code to the intel partition map add-on to allow marking partitions as active after they are created. I did not manage to get it to save the changes to disk yet.

I expected this would be a simple change but it turned out to require more work than I expected (and I'm doing other things as well this week). I hope to have a larger revised patch soon (or maybe I'll manage to separate the changes in different commits).

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>