POSIX: make readv() and writev() conform to IEEE Std 1003.1-2001
The standardized version of readv() and writev() take an int as the third
parameter. Arguably a size_t makes more sense, but the standardization bodies
decided otherwise.

This change fixes the signatures for these functions. This means that it will
break ABI compatibility.

The non-standard functions of readv_pos() and writev_pos() are unchanged. Also
the internal functions in libroot.so, the kernel and the drivers will not
change.

The new behaviour is that when a `count` of less than 0 is given, that the
call will fail with the B_BAD_DATA error.

Change-Id: I949c8ed67dbc0b4e209768cbdee554c929fc242e