POSIX: make readv() and writev() conform to IEEE Std 1003.1-2001
The standardized version of readv() and writev() take an int as the third
parameter. Arguably a size_t makes more sense, but the standardization bodies
decided otherwise.

The non-standard functions of readv_pos() and writev_pos() have been updated
for consistency. The internal functions in libroot.so, the kernel and the
drivers will not change.

The ABI will not change, even though on 64 bit machines the size of the count
parameter will change from 8 to 4 bytes.

The actual use will be slightly different. Like with the size_t argument type,
it will not be possible to give a count lower than 0. If the value is less than
0, then the B_BAD_DATA error will be set. A new change is that the readv/writev
functions (and their nonstandard variants) now respect the IOV_MAX limit from
<limits.h>. Passing in a larger count than IOV_MAX, will set the B_BAD_DATA
error.

Change-Id: I949c8ed67dbc0b4e209768cbdee554c929fc242e