Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> While at it, I would also change readv_pos and writev_pos.

I am sort of on the fence on this one (but could go either way). On the one hand I think that Be got it right by using size_t, since we are describing the size of an array. If I combine that with the fact that readv_pos() and writev_pos() are nonstandard, and the fact that the rest of the internals operate on size_t, I am inclined just to stick with it.

The other side of the argument is that the size must be between 0 and 1024 anyway, and both types are able to express that, so the consistency argument becomes more compelling.

Patch-set: 1
Reviewer: Gerrit User 1000012 <1000012@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>