Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Symbol versions are needed to keep backward compatibility.
> 
> When taking Adrien's input into account: if we only adhere to strict ABI compliance on 32 bit gcc2 systems, and this change is strictly compliant there, then we may just 'break' 64 bit.
> 
> Having said that, if this breaks a large number of binaries, including some fundamentals like coreutils and/or bash, then that's bad. 

I think it wouldn't break anything even on 64bit.

Before we loaded a 64bit value in a 64bit register. With this change we load a 32bit value in that same register, and only use these 32 bits. 

If you have an old app that loads the register with a 64bit value, you can still use only 32bits of it and you will be just fine.

If I missed something and it breaks an application, we can always revert.

Patch-set: 1