Update patch set 5
Patch Set 5:

(1 comment)

> Patch Set 4: Code-Review-2
> 
> > Patch Set 3:
> > 
> > Required changes to locale kit added in https://review.haiku-os.org/c/haiku/+/3172
> 
> It doesn't seem to work, perhaps because BCatalog::SetTo calls MutableLocaleRoster::Default()->UnloadCatalog() before loading requested one. Could you look into it, if my understanding is correct?

I don't follow your reasoning here.

What will happen, as far as I can see:

- The constructor sets fCatalogData = NULL
- Then it calls SetTo(signature)
- This calls UnloadCatalog(NULL), which will do nothing and return B_BAD_VALUE
- Then it will call LoadCatalog

Am I missing something?

One thing I'm not sure about is the construction of the catalogs as static global objects. Maybe it will result in them being constructed too early in the application lifecycle and the locale roster wouldn't be ready yet. In theory, this should be fine (it's the point of having a MutableLocaleRoster->Default() instead of a global variable like, say, be_app), but I'm not sure if everything is done correctly.

One way to test this is to move the static variables inside the functions where they are used. This way, they will be initialized on the first call to the function, which will be later in the app life, and maybe it will help.

Patch-set: 5