IK: Add support for shortcuts without Command modifier
* Introduce a B_NO_COMMAND_KEY modifier that is used when a shortcut
  has no Command modifier, otherwise B_COMMAND_KEY modifier is added.

This is different conceptually from B_NO_MODIFIERS in that it allows
other modifiers such as B_CONTROL_KEY B_SHIFT_KEY and B_OPTION_KEY
to be added to a shortcut -- just not B_COMMAND_KEY. This additional
option is required to remove B_COMMAND_KEY because the Command modifier
role is assumed to be present on all shortcuts. This allows us to break
this assumption removing the Command role modifier without breaking
apps written before this change and assume B_COMMAND_KEY to be present
even if it has not been explictly added to the shortcut.

* Added function key shortcut support. Function keys are encoded as
  B_FUNCTION_KEY_BASE + rawKey. Unicode Private Use Area is used so
  it should not collide with existing symbols.
* BMenuItem use `char` type for shortcut key, so it can't fit function
  keys. SetShortcutEx is introduced for workaround, but
  something better probably should be considered for public interface.
* Use switch insted of array table to improve readability and
  extensibility.
* Added "ShortcutTest" test program.

Fixes #7078

Change-Id: I8b8ec4273ac0c02cb7a8e851dd327d0fcdfa5998