Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> (5 comments)
> 
> From the spec: "EINVAL:  The iovcnt argument was less than or equal to 0, or greater than {IOV_MAX}."

I got confused, because the linux man page choose to accept 0 and just return a value of 0 bytes read if you pass a count of 0. [1]

Then I double checked the spec, and it says it _may_ fail if the count is 0 (but it does not have to).[2] So I understand that this behavior is probably allowed. Checking the upper bound (IOV_MAX) is also a 'may fail', so I am actually considering removing that check as well, to keep the API as compatible as possible with the previous version. What do you think?

[1] https://man7.org/linux/man-pages/man2/read.2.html
[2] https://pubs.opengroup.org/onlinepubs/009695399/functions/readv.html

Patch-set: 2
Reviewer: Gerrit User 1000012 <1000012@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>