Update patch set 1
Patch Set 1:

What does "honor" mean in this context? I know that the rpaths of a lot of the packages that are built by the haikuporter tool are broken, because everything happens in chroots. Will it still fall back to default behavior?

And also: is there a specific issue you are trying to fix?

Patch-set: 1
CC: Gerrit User 1000012 <1000012@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>