Anyboot CD detection, remove nr of partitions check
When booting from floppy image (BOOT_VOLUME_BOOTED_FROM_IMAGE)
we don't know the partition offset, and the number of partitions
are different for 32 bit and 64 bit anyboot cds. So don't check
for number of partitions.

This should only happen when BOOT_VOLUME_BOOTED_FROM_IMAGE
is set. Its intention is to signal if boot was from a floppy
image, so most boot platforms should not set it. It seems they
do, that probably needs a fix of its own.

Change-Id: I6b60a997b72b9928555840358257e15ff417fd42