Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> IMHO we should just double the size of the existing font on large screens.

Why? Using an higher resolution font will give nicer results and the code requires a lot less changes to do that.

I agree with Axel, both fonts should be included and the larger one picked automatically on high resolution displays. I think it shouldn't be too hard to do this in frame_buffer_console.cpp: replace CHAR_WIDTH and CHAR_HEIGHT with values set at init, and use a pointer to one of the two fonts. Having this hidden by an undocimented #define as it is now does not really help.

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>