Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > Why is this code enabled unconditionally anyway? We use anyboot only on x86/64, right?
> 
> No idea, I just fix bugs in the current impl.

We use anyboot on both x86 and x86_64.

However I am surprised that this code can be used in EFI case at all.

From what I understand, the idea is that normally, we are not "booted from image". We are booted from a partition and in that case, the bootloader will tell us which partition it is. There is an exception from this when booting from a CD in BIOS mode. This is because in this (legacy) booting method, we boot from a floppy disk image which doesn't know about the partition.

But in EFI case, BOOT_VOLUME_BOOTED_FROM_IMAGE is set whenever the boot volume is read-only. I'm not sure it should be needed in this case. Even if we are on read-only media, I think the bootloader will already have a quite good idea which partition should be booted.

Anyways...

In this case, and only in this case, the kernel has to guess which partition to start from. So it just looks for something that looks similar to our boot CD layout. Indeed this logic is hardcoded (in many ways). It could do a smarter check, for example accept GPT partitions and check their UUID (that would be a lot more reliable if available and set to a specific value).

Patch-set: 2
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>