Update patch set 2
Patch Set 2:

> > But in EFI case, BOOT_VOLUME_BOOTED_FROM_IMAGE is set whenever the boot volume is read-only. I'm not sure it should be needed in this case. Even if we are on read-only media, I think the bootloader will already have a quite good idea which partition should be booted.
> 
> That sounds like a much better solution. BOOT_VOLUME_BOOTED_FROM_IMAGE really should have some docs about what it is intended for. I don't think I added it, and I've never understood its meaning.

Yes, it makes sense if you know the process of booting from CD in a BIOS based environment, but if you don't know about that you can't really understand it from the name of the constant.

And I think no one really looked into what it does and just tried to behave similarly to the BIOS loader (not only EFI, but also Amiga and Atari bootloaders appear to set this).

Your fix is still correct, so we could still merge this (as a short term fix) but a larger change is needed here.

Patch-set: 2