Update patch set 2
Patch Set 2:

> Even for 32 bit images?

No, after checking, there is no EFI loader in 32bit images.

So maybe we could relax the check here for now and use <= 2 ?

The real solution seems to be to have the anyboot CD use a GPT partition table and use partition UUIDs instead of this guessing.

Patch-set: 2