Update patch set 5
Patch Set 5:

> Patch Set 4:
> 
> I guess the -2 review needs to be removed for it to merged. I thought it just needed rebase.

Yes, in our current config -2 reviews are "sticky" but can be removed by the person who put them, or, in case they are not available, by one of the Gerrit admins.

It is recommended to use only the -1 review if you disagree with something in the current version of a patch but don't want to completely block it.

Patch-set: 5