Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > I very very very strongly disagree. :(
> 
> Why? All code should be tested before merging. We cannot hope to catch all possible scenarios, of course, that's what the nightlies are for; but we should try to hit all the major spots, and avoid causing regressions if possible.
> 
> On this one, that would include: (booting from CD, booting from USB, booting from SATA) on (recent AMD system, recent Intel system, older AMD system, older Intel system, QEMU) with (BIOS loader, EFI loader). That's 5*3*2=30 configurations, yes, but we have people who can help with that, you needn't do it all on your own.

You VETO this change ON YOUR OWN, not because of the code or knowing about this area. And UEFI is a standard that acts same regardless of HW. And master is where we do development. You are changing the rules. I'd get if you had an opinion and -1 it.

Patch-set: 1
Reviewer: Gerrit User 1000021 <1000021@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>