Update patch set 1
Patch Set 1: Code-Review+2

> Patch Set 1:
> 
> > I very very very strongly disagree. :(
> 
> Why? All code should be tested before merging. We cannot hope to catch all possible scenarios, of course, that's what the nightlies are for; but we should try to hit all the major spots, and avoid causing regressions if possible.
> 
> On this one, that would include: (booting from CD, booting from USB, booting from SATA) on (recent AMD system, recent Intel system, older AMD system, older Intel system, QEMU) with (BIOS loader, EFI loader). That's 5*3*2=30 configurations, yes, but we have people who can help with that, you needn't do it all on your own.

This change only affects read-only media because we replace a "IsReadOnly" with "false".

So, booting from USB and SATA are out.

Also, we know the "false" codepath works because it is used in all other cases. And we also know that "booted from image" means "booted from floppy image" as explained in the commit message, and that is simply never the case with EFI (it is a BIOS only legacy thing).

So there is really no reason to put a -2 here. This change is correct and it should be merged.

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=+2