Update patch set 1
Patch Set 1:

> Yes, but we should not merge changes without either (1) a clear benefit, i.e. they fix something, and (2) at least some amount of vetting they do not break anything.

This change fixes the EFI loader incorrectly saying to the kernel that it booted from a floppy image when this is not the case. I don't see how passing obviously invalid parameters to the kernel can be a good thing. So let's remove this.

The other change you have linked made the EFI bootloader do this in all cases, making this even more wrong and breaking more things.

I don't understand why you go straight into -2. It is fair to assume that the change was minimally tested (that is, it boots on at least one machine in the relevant conditions - here, EFI booting from a CD), and if it wasn't, we can blame tqh for it and revert the change. That's fine. You can ask about it, point out more testing scenarios that can be relevant, and you don't need to veto the change for these reasons.

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>